requirements.txt
streamlit==1.32.0
pandas==2.1.4
numpy==1.24.3
requests==2.31.0
fpdf==1.7.2
geopy==2.4.0
earthengine-api==0.1.382
folium==0.14.0
Pillow==10.2.0
python-dotenv==1.0.1
google-generativeai==0.3.2
